Add valgrind submodule and cherry-pick history
Initially memtrace was developed as patches on top of valgrind, but
now this became a maintenance burden. Having valgrind as a submodule
is much nicer, but implementing this the right way will result in an
ugly commit with over 9000 deletions. Therefore be bold and pretend
things were this way since day 1.